+++ src/Service/ReservationService.java
@@ -6,34 +6,19 @@ import Data.Reservation;
 import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Scanner;
 import java.util.stream.Collectors;
 
 public class ReservationService {
     private List<Reservation> reservations;
     private CoworkingSpaceService spaceService;
-    private Scanner scanner;
 
-    public ReservationService(CoworkingSpaceService spaceService, Scanner scanner) {
+    public ReservationService(CoworkingSpaceService spaceService) {
         this.reservations = new ArrayList<>();
         this.spaceService = spaceService;
-        this.scanner = scanner;
         loadReservationsFromFile();
     }
 
-    public void addReservation() {
-        System.out.print("Enter your name: ");
-        scanner.nextLine();
-        String userName = scanner.nextLine();
-        System.out.print("Enter date (YYYY-MM-DD): ");
-        String date = scanner.nextLine();
-        System.out.print("Enter start time (HH:MM): ");
-        String startTime = scanner.nextLine();
-        System.out.print("Enter end time (HH:MM): ");
-        String endTime = scanner.nextLine();
-        System.out.print("Enter the ID of the space you want to book: ");
-        int spaceId = scanner.nextInt();
-
+    public void addReservation(String userName, String date, String startTime, String endTime, int spaceId) {
         CoworkingSpace space = spaceService.getSpaceById(spaceId);
         if (space == null || !space.isAvailable()) {
             System.out.println("Space not available or does not exist.");
@@ -47,40 +32,22 @@ public class ReservationService {
         }
     }
 
-    public void cancelReservation() {
-        System.out.print("Enter your name: ");
-        scanner.nextLine();
-        String userName = scanner.nextLine();
+    public void cancelReservation(int reservationId) {
+        Reservation reservation = getReservationById(reservationId);
+        if (reservation != null) {
+            reservation.getSpace().setAvailable(true);
+            reservations.remove(reservation);
+            System.out.println("Reservation canceled successfully!");
+            saveReservationsToFile();
+        } else {
+            System.out.println("Reservation not found.");
+        }
+    }
 
-        List<Reservation> userReservations = reservations.stream()
+    public List<Reservation> getReservationsByUser(String userName) {
+        return reservations.stream()
                 .filter(reservation -> reservation.getUserName().equals(userName))
                 .collect(Collectors.toList());
-
-        if (userReservations.isEmpty()) {
-            System.out.println("No reservations found for " + userName);
-        } else {
-            for (Reservation reservation : userReservations) {
-                System.out.println("Reservation ID: " + reservation.getId());
-                System.out.println("Date: " + reservation.getDate());
-                System.out.println("Time: " + reservation.getStartTime() + " - " + reservation.getEndTime());
-                System.out.println("Space ID: " + reservation.getSpace().getId());
-                System.out.println("------------------------");
-            }
-
-            System.out.print("Enter the ID of the reservation to cancel (or 0 to go back): ");
-            int reservationId = scanner.nextInt();
-            if (reservationId != 0) {
-                Reservation reservation = getReservationById(reservationId);
-                if (reservation != null) {
-                    reservation.getSpace().setAvailable(true);
-                    reservations.remove(reservation);
-                    System.out.println("Reservation canceled successfully!");
-                    saveReservationsToFile();
-                } else {
-                    System.out.println("Reservation not found.");
-                }
-            }
-        }
     }
 
     public void displayAllReservations() {
